fix refresh button to work
